+++ src/Angor/Avalonia/AngorApp/Sections/Founder/Details/IInvestmentViewModel.cs
@@ -1,0 +1,13 @@
+using Zafiro.UI.Commands;
+
+namespace AngorApp.Sections.Founder.Details;
+
+public interface IInvestmentViewModel
+{
+    public IAmountUI Amount { get; }
+    public string InvestorNostrPubKey { get; }
+    public DateTimeOffset Created { get; }
+    public IEnhancedCommand<Unit, Maybe<Result<bool>>> Approve { get; }
+    public InvestmentStatus Status { get; }
+    public bool CanApprove => Status == InvestmentStatus.Pending;
+}
