+++ src/Angor/Avalonia/AngorApp/Sections/Founder/Details/InvestmentStatus.cs
@@ -1,0 +1,9 @@
+namespace AngorApp.Sections.Founder.Details;
+
+public enum InvestmentStatus
+{
+    None,
+    Pending,
+    Approved,
+    Invested,
+}
